pyproject.toml
[project]
name = "rl-loop-demo"
version = "1.0.0"
description = "CUDA IPC 跨进程 GPU 显存共享 RL 演示项目"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cupy-cuda12x>=13.0.0",
    "numpy>=2.0.0",
    # NVIDIA CUDA 运行时库 (解决 libnvrtc.so.12 等缺失问题)
    "nvidia-cuda-nvrtc-cu12",
    "nvidia-cuda-runtime-cu12",
    "nvidia-curand-cu12",
    "nvidia-cublas-cu12",
    "nvidia-cufft-cu12",
    "nvidia-cusolver-cu12",
    "nvidia-cusparse-cu12",
]
